Bluetooth: Add maxListCount param in getfolderlisting.
A special case, where MCE APP send the folderlisting
with maxlistcount=0, startoffset=5, As DUT(MCE role)
does not include maxListcount param in folderlsiting,
when it's value set to 0 by MCE APP. So as per the spec,
MSE assume  maxListCount = 1024 for folderlsiting & MSE
response as the bad request, when startoffset is Out of
range. Adding MaxListCount in folderListing to
solve this issue, when it's value set to 0.

Change-Id: I03a08c18f1b45a87a33077dbeb18991ac3170973
CRs-Fixed: 791193